Treat tile_x and tile_y as projection system units.
Previously, tile_x and tile_y were assumed to be the number of pixels per tile. However, this makes thinking about and working with tile specs more difficult. Given that the shift_* and pixel_* values are in terms of projection system coordinates, this change makes this set of values consistent.