Use lcmap.config.cassaforte for connect params
Several components across several projects need to create connections to Cassandra. The connection options are a nested map that aren't amenable to the way values are organized in an INI file. Ultimately, using a schema to coerce these values into the right shapes and types will probably make sense. However, this is step in the right direction.

Components can use the provided function to prepare basic properties (port, connection limits, credentials). More complex properties (e.g., retry, load-balancing, replication) need to be configured manually. It might be best to use the Cassaforte policy binding functions though in the context where this matters in the mean time.