Save tiles with and without a keyspace or table provided.
If needed, retrieve a tile-spec (as is the case when saving a tile via a REST API call).

Still provide a way for ingest to save tiles without looking up the target keyspace and table for the tile data.

Removed base64 decoding call, this was a temporary hack, something the REST API should perform.